Add code

+++ backend/resolvers/Mutation.js
@@ -42,15 +42,33 @@ const Mutation = {
 
     async deleteTeam(parent, { name }, { db, pubsub }, info){
         if(!name) throw new Error("Missing team name in mutation deleteTeam");
-        const toDelete = await db.TeamDataModel.deleteOne({team: name});
-        if(toDelete) return true;
-        else return false;
+        const existing = await db.TeamDataModel.findOne({team: name});
+
+        if(existing) {
+            pubsub.publish(`deleteTeam`, {
+                deleteTeam: existing,
+            });
+
+            const matchFound = await db.MatchModel.find({$or: [{team_1: name}, {team_2: name}]});
+            const n = matchFound.length;
+            for(let i = 0; i < n; i++){
+                const matchToDelete = matchFound[i];
+                await db.MatchModel.deleteOne(matchToDelete);
+                pubsub.publish('allMatch', {
+                    allMatch: {
+                        mutation: "CREATED",
+                        match: matchToDelete,
+                    },
+                });
+            }
+
+            await db.TeamDataModel.deleteOne({team: name});
+            return true;
+        }else return false;
     },
 
     async updateTime(parent, { name, time }, { db, pubsub }, info){
         console.log(name,time);
-        if(!name) throw new Error("Missing team name in mutation updateTime");
-        if(!time) throw new Error("Missing team time in mutation updateTime");
         const existing = await db.TeamDataModel.findOneAndUpdate({team: name}, {time: time}, {new: true});
         
         pubsub.publish(`team ${name}`, {
@@ -125,7 +143,10 @@ const Mutation = {
                                             // console.log(matchdata);
                                             matchdata.save();
                                             pubsub.publish('allMatch', {
-                                                allMatch: matchdata,
+                                                allMatch: {
+                                                    mutation: "CREATED",
+                                                    match: matchdata,
+                                                },
                                             });
                                             pubsub.publish(`team ${teamData.team} match`, {
                                                 teamMatch: matchdata,
